Make hcAction hook & use in IG & inst tlbarActions
